feat: Add Mesh class and related methods
This commit adds the Mesh class, which is used for mesh tesselation in 2D material domains. The class can generate triangular, quad, and polygonal finite elements. It works seamlessly with SDF.m.

The main usage of the Mesh class is to convert an SDF (signed distance function) class to a MESH class.

Some significant changes include:
- Adding the Mesh.m file, which contains the implementation of the