feat: Add support for generating element node degrees of freedom
This commit adds support for calculating the number of degrees of freedom per element node in the Mesh class. The `ElemNDof` property is now calculated as 3 times the length of each row in the `f` matrix. This change allows for more accurate representation and manipulation of element data.

Also, update meshfield.m documentation

The documentation for the `meshfield` function has been updated to provide a detailed explanation of its usage, inputs, and outputs. Examples have also been added to demonstrate how to use the function with a mesh object and vector field values.